app/requirements.txt
flask
flask_login
flask_migrate
flask_wtf
flask_sqlalchemy
flask_bcrypt
gunicorn
pillow
keras
mtcnn
flask_mail
flask_cors
flask_socketio
sqlalchemy-datatables
mxnet
sklearn
eventlet
flask_cors
flask_socketio
opencv-python
Flask-Migrate
flask_babel
tensorflow
scikit-image
Flask-JWT
SQLAlchemy-serializer
redis
rq
email_validator
apscheduler
pyarmor